Remove sub fold cross validation
