+++ Tests/ML/visualizers/test_plot_cross_validation.py
@@ -76,19 +76,17 @@ def download_metrics(config: PlotCrossValidationConfig) -> \
     return dataframes, root_folder
 
 
-def create_run_result_file_list(config: PlotCrossValidationConfig, folder: str,
-                                perform_sub_fold_cross_validation: bool = False) -> List[RunResultFiles]:
+def create_run_result_file_list(config: PlotCrossValidationConfig, folder: str) -> List[RunResultFiles]:
     """
     Creates a list of input files for cross validation analysis, from files stored inside of the test data folder.
     :param config: The overall cross validation config
     :param folder: The folder to read from, inside of test_data/plot_cross_validation.
-    :param perform_sub_fold_cross_validation: If True then create input files for sub fold cross validation analysis.
     :return:
     """
     full_folder = full_ml_test_data_path("plot_cross_validation") / folder
     files: List[RunResultFiles] = []
     previous_dataset_file = None
-    for split in ["0", "1", "1", "1"] if perform_sub_fold_cross_validation else ["0", "1"]:
+    for split in ["0", "1"]:
         for mode in config.execution_modes_to_download():
             metrics_file = full_folder / split / mode.value / METRICS_FILE_NAME
             dataset_file: Optional[Path] = full_folder / split / DATASET_CSV_FILE_NAME
@@ -131,7 +129,7 @@ def test_metrics_preparation_for_segmentation(test_config_ensemble: PlotCrossVal
         pd.testing.assert_frame_equal(expected_df, actual_df, check_like=True, check_dtype=False)
 
 
-def load_result_files_for_classification(perform_sub_fold_cross_validation: bool = False) -> \
+def load_result_files_for_classification() -> \
         Tuple[List[RunResultFiles], PlotCrossValidationConfig]:
     plotting_config = PlotCrossValidationConfig(
         run_recovery_id="local_branch:HD_cfff5ceb-a227-41d6-a23c-0ebbc33b6301",
@@ -139,24 +137,21 @@ def load_result_files_for_classification(perform_sub_fold_cross_validation: bool
         model_category=ModelCategory.Classification
     )
     files = create_run_result_file_list(config=plotting_config,
-                                        folder="HD_cfff5ceb-a227-41d6-a23c-0ebbc33b6301",
-                                        perform_sub_fold_cross_validation=perform_sub_fold_cross_validation)
+                                        folder="HD_cfff5ceb-a227-41d6-a23c-0ebbc33b6301")
     return files, plotting_config
 
 
-@pytest.mark.parametrize("perform_sub_fold_cross_validation", [True, False])
-def test_metrics_preparation_for_classification(perform_sub_fold_cross_validation: bool) -> None:
+def test_metrics_preparation_for_classification() -> None:
     """
     Test if metrics from classification models can be loaded and prepared. The files in question are checked in,
     and were downloaded from a run on AzureML.
     """
-    files, plotting_config = load_result_files_for_classification(perform_sub_fold_cross_validation)
+    files, plotting_config = load_result_files_for_classification()
     downloaded_metrics = load_dataframes(files, plotting_config)
     assert ModelExecutionMode.TEST not in downloaded_metrics
     metrics = downloaded_metrics[ModelExecutionMode.VAL]
     assert metrics is not None
-    expected_metrics_file = "metrics_preparation_for_sub_fold_classification_VAL.csv" \
-        if perform_sub_fold_cross_validation else "metrics_preparation_for_classification_VAL.csv"
+    expected_metrics_file = "metrics_preparation_for_classification_VAL.csv"
     expected_df_csv = full_ml_test_data_path("plot_cross_validation") / expected_metrics_file
     metrics = metrics.sort_values(list(metrics.columns), ascending=True).reset_index(drop=True)
     # To write new test results:
